requirements.txt
fastapi
uvicorn
python-dotenv
httpx
datasets
pydantic
numpy
pandas
tqdm
pytest
pytest-asyncio
google-genai
a2a-sdk
starlette
